refactor: extract TranscriptStoreRecoveryEvent.swift from TranscriptStore.swift
Extracts the TranscriptStoreRecoveryEvent struct (18 lines) from
TranscriptStore.swift into its own file. The public struct describes
recovery events emitted after transcript load failures.

Byte-identical move; SHA verified against origin/main:TranscriptStore.swift
lines 554-571.

Reduces TranscriptStore.swift from 571 to 552 lines.

Tests: 667 (unchanged).
