+++ Sources/BugNarrator/Services/TranscriptStore.swift
@@ -550,22 +550,3 @@ private struct TranscriptStoreIndex: Codable {
         self.sessionIDs = sessionIDs ?? entries.map(\.id)
     }
 }
-
-struct TranscriptStoreRecoveryEvent: Equatable {
-    enum Source: Equatable {
-        case backup
-        case failed
-    }
-
-    let source: Source
-    let recoveredSessionCount: Int
-
-    var userMessage: String {
-        switch source {
-        case .backup:
-            return "Session history was recovered from the local backup. \(recoveredSessionCount) session\(recoveredSessionCount == 1 ? "" : "s") restored."
-        case .failed:
-            return "Session history could not be read from the primary or backup store. A new empty library was opened."
-        }
-    }
-}
